==> 01 flight delay prediction/02_preprocess_flight_history_departures.Rmd <==
---
title: "02 Preprocess Flight History Departures"
author: "Nils Hellwig"
date: "5/16/2023"
output: html_document
---

# Notebook: Create Dataset with Departures

This notebook is used to preprocess the downloaded .html files of the departure flights to assemble them into a neatly formatted dataframe. This script takes about an hour to run (at least with my M1 MacBook).

## Load Packages

```{r}
library(rvest)
library(tidyverse)
library(stringr)
library(xml2)
source("date_string_formatter.R")
library(lubridate)
```

## Define Settings

```{r}
folder_path_departures <- "../datasets/raw_departures_dataset"
```

## Code

### Load HTML-files

```{r}
# Load the filepaths of all the .html-documents
html_documents <- list.files(path = folder_path_departures, pattern = "\\.html$", full.names = TRUE)
head(html_documents)
```

### Store collected data to dataframe departure_data

```{r}
departure_data <- data.frame(
  departure_scheduled_datetime_local = character(),
  departure_scheduled_time_local_timezone = character(),
  departure_scheduled_datetime_utc = character(),
  flight_status = character(),
  airline_name = character(),
  flight_number = character(),
  flight_number_IATA = character(),
  airline_code = character(),
  airline_code_IATA = character(),
  airline_country = character(),
  destination_country = character(),
  destination_name = character(),
  destination_IATA = character(),
  destination_ICAO = character(),
  departure_actual_datetime_local = character(),
  delay_departure_min = numeric(),
  delay_arrival_min = numeric(),
  actual_arrival_datetime_local = character(),
  flight_duration_time_min = numeric()
)
```

```{r}
# Load each .html-document
for (i in seq_along(html_documents)) {
  doc_path <- html_documents[i]
  doc <- read_html(doc_path)
  
  # The flight data for the flights are located within the '<tbody>' element. Each flight is 
  # within a '<tr>' element that has either the class "bg-gray-50" or "bg-white". 
  # The class was added as a condition, because there can be other '<tr>' elements within the <tbody> element, but they do not contain flight data.
  extracted_data <- doc %>%
    html_nodes("tbody tr.bg-gray-50, tbody tr.bg-white")
  
  for (flight_element in extracted_data) {
      # Let's start by storing all innerHTML of the child elements of a flight's flight_element.
      # In each child element there is metadata about a flight. I hoped that each child element would have an id/class with 
      # which to extract the corresponding metadata, but unfortunately this was not the case. However, I noticed that 
      # the index or the order of the child elements within a flight_element is sufficient to know which child element contains which information. 
      # For example, child_elements[2] always contains the airline code and airline name for each flight. airline code 
      # and airline name then only have to be extracted using regular expressions.
      child_elements <- flight_element %>%
        html_children() %>%
        html_text() %>%
        trimws() # remove whitespaces
      
      # The image of a flight is used to identify the destination country / airline country      
      flag_imgs <- flight_element %>%
         html_nodes("img[alt*='Flag for']")

      # It happens that an airline can not be assigned to a country so that there's no flag icon.
      # Example: Flight to Liege (Belgium), ATRAN-Aviatrans Cargo Airlines, see here: https://www.flightera.net/en/airport/Munich/EDDM/departure/2017-12-13%2022_00?
      if (length(flag_imgs) == 1) {
         airline_country <- NULL
         destination_country <- str_extract(html_attr(flag_imgs[1], "alt"), "(?<=Flag for ).*")
      } else {
         airline_country <- str_extract(html_attr(flag_imgs[1], "alt"), "(?<=Flag for ).*")
         destination_country <- str_extract(html_attr(flag_imgs[2], "alt"), "(?<=Flag for ).*")
      }
     
      # There are either one or two flight numbers. The flight number is located before the first two \n\n inside child_elements[2]
      flight_number <- gsub("\\n\\n.*", "", child_elements[2])
      flight_number <- str_extract_all(flight_number, "[A-Z\\d]+")
      
      airline <- strsplit(child_elements[2], "\n")[[1]]
      airline <- tail(airline, n = 1)
    
      airline <- strsplit(airline, "/")[[1]]
      airline_code <- tail(strsplit(airline[1], " ")[[1]], n = 1)
      airline_code_IATA <- airline[2]
      airline_name <- sub("\\s\\S+$", "", airline)[1]
      
      departure_scheduled_date_local <- convert_date_string(strsplit(child_elements[1], "\n")[[1]][1])
      flight_status <- tail(gsub("\\s+", "",strsplit(child_elements[1], "\n")[[1]]), n = 1)
 
      # Although we know that the local timezone of munich is CEST, I decided to extract that data too because 
      # one might want to use the script for another airport
      if (year(as.Date(departure_scheduled_date_local))==2023) {
        departure_scheduled_time_local <- strsplit(child_elements[1], "\n")[[1]][7]
        departure_scheduled_time_local_timezone <- strsplit(child_elements[1], "\n")[[1]][8]
      } else {
        departure_scheduled_time_local <- strsplit(child_elements[1], "\n")[[1]][8]
        departure_scheduled_time_local_timezone <- strsplit(child_elements[1], "\n")[[1]][9]
      }

      departure_scheduled_datetime_local <- paste(departure_scheduled_date_local, departure_scheduled_time_local)
      departure_scheduled_datetime_utc <- format(with_tz(ymd_hm(departure_scheduled_datetime_local, tz = "CEST"), "UTC"), "%Y-%m-%d %H:%M")
     
      destination_name <- sub(" \\(.*", "", child_elements[3])
      
      # str_match returns the second column of these matrices
      flight_code_IATA <- str_match(child_elements[3], "\\(([A-Z]{3}) /")[, 2]
      flight_code_ICAO <- str_match(child_elements[3], " / ([A-Z]{4})\\)")[, 2]
      
      # Information about the delay is only available if the flight has landed and has not been cancelled or for the case that the status is unknown.
      if (flight_status == "Landed") {
        delay_departure_split <- strsplit(child_elements[5], "\n")[[1]]
        # extract_minutes_from_delay_string() extracts the number of minutes from a string like "23min late" or "1h 15min late"
        delay_departure_min <- extract_minutes_from_delay_string(delay_departure_split)
        departure_actual_datetime_local <- format(ymd_hm(departure_scheduled_datetime_local) + minutes(delay_departure_min), "%Y-%m-%d %H:%M")
        delay_arrival_split <- strsplit(child_elements[6], "\n")[[1]]
        delay_arrival_min <- extract_minutes_from_delay_string(delay_arrival_split)
        
        flight_duration_time_min <- extract_minutes_from_delay_string(child_elements[7])
        actual_arrival_datetime_local <- format(ymd_hm(departure_scheduled_datetime_local) + minutes(flight_duration_time_min + delay_departure_min), "%Y-%m-%d %H:%M")
      }
      
      
      flight_number_IATA <- flight_number[[1]][2]
      flight_number <- flight_number[[1]][1]
      
      # Finally, we can add all the extracted information regarding a specific flight to our dataframe
      departure_data <- departure_data %>%
        add_row(
          departure_scheduled_datetime_local = departure_scheduled_datetime_local,
          departure_scheduled_time_local_timezone = departure_scheduled_time_local_timezone,
          departure_scheduled_datetime_utc = departure_scheduled_datetime_utc,
          airline_code = airline_code,
          airline_code_IATA = airline_code_IATA,
          airline_name = airline_name,
          airline_country = airline_country,
          flight_number = flight_number[[1]][1],
          flight_number_IATA = flight_number_IATA,
          flight_status = flight_status,
          departure_actual_datetime_local = departure_actual_datetime_local,
          destination_name = destination_name,
          destination_IATA = flight_code_IATA,
          destination_ICAO = flight_code_ICAO,
          destination_country = destination_country,
          delay_departure_min = delay_departure_min,
          delay_arrival_min = delay_arrival_min,
          actual_arrival_datetime_local = actual_arrival_datetime_local,
          flight_duration_time_min = flight_duration_time_min
        )
  }
}
```

### Further Preprocessing of the dataset

If there are not at least 30 flights available for an hour on flightera.com, flights for more than the randomly selected hour may be crawled. In rare cases duplicates may occur in case the upcoming hour was also randomly selected for data collection.

```{r}
duplicate_rows <- departure_data[duplicated(departure_data), ]
departure_data <- distinct(departure_data)
head(duplicate_rows)
```

```{r}
# For better data handling, I´ll create separate columns for month, year and hour of the planned departures.
departure_data$departure_scheduled_datetime_local <- as.POSIXct(departure_data$departure_scheduled_datetime_local, format = "%Y-%m-%d %H:%M")
departure_data$departure_scheduled_month_year_local <- format(departure_data$departure_scheduled_datetime_local, "%Y-%m")
departure_data$departure_scheduled_month_local <- format(departure_data$departure_scheduled_datetime_local, "%m")
departure_data$departure_scheduled_year_local <- format(departure_data$departure_scheduled_datetime_local, "%Y")
departure_data$departure_scheduled_hour_local <- format(departure_data$departure_scheduled_datetime_local, format = "%H")

# Setting the language with Sys.setenv is relevant to show the delays correctly depending on the week (names should be in English in the plot)
Sys.setlocale("LC_TIME", "C")
weekday_order <- c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday")
departure_data$departure_scheduled_weekday_local <- factor(weekdays(departure_data$departure_scheduled_datetime_local), levels = weekday_order)
```

It is possible that flight data outside the period I am looking at is listed on flightera, as sometimes data is returned for more than just the requested hour, as already described.

```{r}
departure_data <- departure_data[format(departure_data$departure_scheduled_datetime_local, "%Y-%m") >= "2017-10" & format(departure_data$departure_scheduled_datetime_local, "%Y-%m") <= "2023-05", ]
```


```{r}
head(departure_data)
```

```{r}
write.csv(departure_data, file = "../datasets/departure_dataset.csv", row.names = FALSE)
```
