==> 01 flight delay prediction/03_dataset_exploration.Rmd <==
# Notebook: Explore Datasets

In order to gain a better understanding of what the data collected looks like, important characteristics of the dataset will be presented/visualised.

## Load Packages

```{r}
library(ggplot2)
library(dplyr)
source("get_weather_data.R")
```

## Code

### Load Datasets

First, we need to load the dataset and add some columns that will be used for exploring the dataset

```{r}
# Load departure data
departure_data <- read.csv("../datasets/departure_dataset.csv")
weather_data <- read.csv("../datasets/weather_munich_dataset.csv")
```

For better data handling, I will convert the columns with time data to POSIXct

```{r}
departure_data$departure_scheduled_datetime_local <- as.POSIXct(departure_data$departure_scheduled_datetime_local)
weather_data$six_hours_starting_from_datatime <- as.POSIXct(paste(weather_data$date, weather_data$six_hours_starting_from))
```

Add weather data to departure dataset (using the script from get_weather_data.R)

```{r}
departure_data$six_hours_starting_from_datatime <- as.POSIXct(sapply(departure_data$departure_scheduled_datetime_local, get_weather_data))
```

Exclude unnecessary columns that we don't want to add to the departure_data 

```{r}
merge_cols_weather_data <- setdiff(names(weather_data), c("date", "six_hours_starting_from"))
departure_data <- merge(departure_data, weather_data[, c(merge_cols_weather_data)], by.x = "six_hours_starting_from_datatime", by.y = "six_hours_starting_from_datatime", all.x = TRUE)
head(departure_data)
```

### Explore Overall Departure Data

In a first step, we can explore the dataset and its characteristics. 
For every analysis, it's important to keep in mind, that the dataset is only a radnom sample as I only collected data for four random hours of each day (see 01_... notebooks). For each day between 01.10.2017 and 31.05.2023 flights that took off in Munich were collected from flightera.org.
Now, we can have a look at how many flights have landed, were diverted etc. (occurrences of the various flight statuses)

```{r}
table(departure_data$flight_status)
```

Due to the Covid-19 pandemic, strong differences between the years can be observed here.

```{r, fig.width = 10, fig.height = 20}
flight_status_count <- departure_data %>%
  group_by(departure_scheduled_year_local, departure_scheduled_month_local, flight_status) %>%
  summarise(count = n())

month_labels <- c("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")

ggplot(flight_status_count, aes(x = departure_scheduled_month_local, y = count, fill = flight_status)) +
  geom_bar(stat = "identity", position = "stack") +
  facet_wrap(~departure_scheduled_year_local, ncol = 1) +
  xlab("Month") +
  ylab("# Flights") +
  ggtitle("Distribution of Flight Status by Month (Year-wise)") +
  scale_x_continuous(breaks = 1:12, labels = month_labels)
```

Next, we can look at how many flights were *planned* to take off in the period for which the data was collected for.
Here it can be seen that in 2020, the year that was strongly influenced by the pandemic, there was a decrease in the amount of flights.

```{r}
flight_count_departure <- departure_data %>%
  group_by(departure_scheduled_month_year_local) %>%
  summarise(count = n())

ggplot(flight_count_departure, aes(x = as.Date(paste(departure_scheduled_month_year_local, "01", sep = "-")), y = count)) +
  geom_line() +
  scale_x_date(date_labels = "%b %Y", date_breaks = "1 year") +
  labs(x = "Month", y = "# Flights") +
  ggtitle("Departures from Munich Airport")
```

Next, we can look at which airline's flights are most frequently scheduled to take off from Munich.

```{r}
airlines <- departure_data %>%
  group_by(airline_name) %>%
  summarise(count = n()) %>%
  arrange(-count)

head(airlines, n=20)
```

Of course, this can also be displayed in a bar chart with the percentage share of each airline in the overall dataset. This might be better to show the strong differences between the airlines.

```{r}
top_20_airlines_departure <- departure_data %>%
  count(airline_name) %>%
  mutate(percentage = (n / nrow(departure_data)) * 100) %>%
  arrange(-percentage) %>%
  head(20)

ggplot(top_20_airlines_departure, aes(x = reorder(airline_name, -percentage), y = percentage)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Airline", y = "% of all collected Flights", title = "Most frequent airlines departing from Munich Airport") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

You can also look at which airports are the most frequent destinations among flights departing from Munich

```{r}
destinations <- departure_data %>%
  group_by(destination_IATA, destination_name) %>%
  summarise(count = n()) %>%
  arrange(-count)

head(destinations, n=20)
```

... or by country

```{r}
top_20_destinations_departure <- departure_data %>%
  count(destination_country) %>%
  mutate(percentage = (n / nrow(departure_data)) * 100) %>%
  arrange(-percentage) %>%
  head(20)

ggplot(top_20_destinations_departure, aes(x = reorder(destination_country, -percentage), y = percentage)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Destination (Country)", y = "% All Collected Flights", title = "Most frequent destinations from Munich Airport") +
  # by changing the angle of the labels by 45° it's much easier to read them
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

### Feature Exploration / Prediction Target

Next, I can look at which columns are possible predictors. 

```{r}
delay_departure_data <- departure_data 
```

```{r}
delay_departure_data$departure_scheduled_datetime_local <- as.Date(delay_departure_data$departure_scheduled_datetime_local)
```

Next, we can examine how often and how much delayed flights are taking off from Munich. To analyse delays, we only look at flights that have landed and not flights that have been cancelled, diverted or whose status is unknown etc.. We also need to remove outliers. There seem to be flights with extreme delays, which probably does not correspond to reality but might be due to technical issues.

```{r}
z_scores <- scale(delay_departure_data$delay_departure_min)
z_threshold <- 3
outliers <- abs(z_scores) > z_threshold
outlier_data <- delay_departure_data[outliers, ]
delay_departure_data <- delay_departure_data[!outliers, ]

# Among the collected flights there were some with more than one day of delay (>1440 minutes)
head(outlier_data[order(-outlier_data$delay_departure_min), ])[, c("delay_departure_min", "departure_scheduled_datetime_utc", "destination_name")]
```

```{r}
delay_departure_data <- delay_departure_data[delay_departure_data$flight_status == "Landed", ]
```

#### Prediction target

Between the 0.2 and the 0.8 quantile the slope is fairly constant, while between the 0.8 and the 1.0 quantile, one can see a clear increase. Between the 0.0 and the 0.2 quantile, the slope is at 0 for some time, which can be explained by the fact that many flights have exactly 0 minutes delay.

```{r}
summary(delay_departure_data$delay_departure_min)
```

```{r}
step <- 0.001
category_quantiles <- quantile(delay_departure_data$delay_departure_min, probs = seq(0, 1, by = step))
plot(seq(0, 1, by = step), category_quantiles, type = "l", 
     xlab = "Quantile", ylab = "Delay in minutes", 
     main = "Quantiles for flight delay", 
      ylim = c(-50, 50)
    )
     
grid()
legend("topleft", legend = "Quantile", lty = 1, col = "blue")
```

The goal will eventually be to predict whether a flight has...

* no or little delay 
* a long or very long delay

I decided to split these two groups based on the median. Further notes on the prediction task can also be found in notebook 04.

```{r}
quantiles <- quantile(delay_departure_data$delay_departure_min, probs = c(0.0, 0.5, 1.0))
quantiles
```

#### Feature Exploration

Among other things, you can see, there are some flights with 0 minutes delay (="on time"):

```{r}
minutes <- seq(min(delay_departure_data$delay_departure_min), max(delay_departure_data$delay_departure_min), by = 1)
hist(delay_departure_data$delay_departure_min, breaks = minutes, 
     xlab = "Delay in minutes", ylab = "Frequency", 
     main = "Distribution of Departure Delay in 1-Minute Intervals",
     xlim = c(-20, 100),
     #ylim = c(0, 8000)
     )
```

Looking at the average number of minutes of delay for all months, it is noticeable that in the years that were heavily affected by the Covid 19 pandemic (2020 and 2021), the number of minutes of delay was lower than before and after that time. Overall, it can be observed that there are strong differences between the months.

```{r}
average_delay_destination_departures <- delay_departure_data[delay_departure_data$flight_number == "XQ131", ] %>%
  group_by(departure_scheduled_month_year_local) %>%
  summarise(avg_delay = mean(delay_departure_min))

ggplot(average_delay_destination_departures, aes(x = as.Date(paste(departure_scheduled_month_year_local, "01", sep = "-")), y = avg_delay)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  scale_x_date(date_labels = "%b %Y", date_breaks = "1 year") +
  labs(x = "Month", y = "Average Departure Delay (Minutes)") +
  ggtitle("Average Departure Delay of flights from Munich Airport")
```
As can already be seen in the figure, the average delays were not that high in 2020. Accordingly, the class below the median also occurs rather rarely in 2020 compared to the class that includes examples whose delay_departure_min is above the median. In April 2020, for example, only 5% of the examples belong to this class.

```{r}
quartile_50 <- quantile(delay_departure_data$delay_departure_min, probs = 0.5)

delay_departure_data <- delay_departure_data %>%
  mutate(delay_class = ifelse(delay_departure_min < quartile_50, "earlier, on time or very short delay", "long delay"))

class_counts_each_month_and_year <- delay_departure_data %>%
  group_by(departure_scheduled_month_year_local) %>%
  summarise(count_earlier = sum(delay_class == "earlier, on time or very short delay"),
            count_long = sum(delay_class == "long delay")) %>%
  mutate(percentage_earlier = count_earlier / (count_earlier + count_long),
         percentage_long = count_long / (count_earlier + count_long))

class_counts_each_month_and_year
```

There also seems to be a difference in how much the average delay is depending on the month

```{r}
ggplot(delay_departure_data, aes(x = as.factor(departure_scheduled_month_local), y = delay_departure_min)) +
  geom_boxplot() +
  labs(x = "Month", y = "Delay (Minutes)", title = "Distribution of Delay per Month") +
  coord_cartesian(ylim = c(0, 100))
```

This trend can also be seen if one excludes the two years 2020 and 2021, which were heavily affected by the pandemic.

```{r}
years_no_pandemic <- c(2017, 2018, 2019, 2020, 2023)

ggplot(delay_departure_data %>% filter(departure_scheduled_year_local %in% years_no_pandemic), aes(x = as.factor(departure_scheduled_month_local), y = delay_departure_min)) +
  geom_boxplot() +
  labs(x = "Month", y = "Delay (Minutes)", title = "Distribution of Delay per Month [without 2020 and 2021]") +
  coord_cartesian(ylim = c(0, 100))
```

In addition, there also seem to be differences depending on the time of day (hour) as to how long the delay is when flights depart. 

```{r}
ggplot(delay_departure_data, aes(x = as.factor(departure_scheduled_hour_local), y = delay_departure_min)) +
  geom_boxplot() +
  labs(x = "Hour", y = "Delay (Minutes)", title = "Distribution of delay at different times of the day") +
  coord_cartesian(ylim = c(0, 50))
```
There also seem to be differences depending on the day of the week, with the average delay being lowest on Tuesdays and highest on Fridays.

```{r}
delay_departure_data$departure_scheduled_weekday_local <- factor(delay_departure_data$departure_scheduled_weekday_local,
                                                               levels = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"))

ggplot(delay_departure_data, aes(x = departure_scheduled_weekday_local, y = delay_departure_min)) +
  geom_boxplot() +
  labs(x = "Weekday", y = "Delay (Minutes)", title = "Distribution of delay at different weekdays") +
  coord_cartesian(ylim = c(0, 50))
```

There also seem to be differences between the airlines as to how long the delay is. For this we can look at the airlines that have at least one flight a day on average. 

```{r}
top_airlines_delay <- delay_departure_data %>%
  group_by(airline_name, airline_country) %>%
  summarise(delay_mean = mean(delay_departure_min), delay_median = median(delay_departure_min), count = n()) %>%
  filter(count >= 365) %>%
  arrange(-delay_mean)

head(top_airlines_delay, n=20)
```

```{r}
top_flights_delay <- delay_departure_data %>%
  group_by(flight_number) %>%
  summarise(delay_mean = mean(delay_departure_min), delay_median = median(delay_departure_min), count = n()) %>%
  filter(count >= 365) %>%
  arrange(-delay_mean)

head(top_flights_delay, n = 20)
```


Next, we can look at whether the destination has an influence. If we look again at destinations that are flown to at least once a day on average, we can see that there are differences.

```{r}
top_destination_delay <- delay_departure_data %>%
  group_by(destination_IATA, destination_name) %>%
  summarise(delay_mean = mean(delay_departure_min), delay_median = median(delay_departure_min), count = n()) %>%
  filter(count >= 365) %>%
  arrange(-delay_mean)

head(top_destination_delay, n=20)
```

##### Weather data

Let's have a look at how the weather might have an influence on delays. Let's have a look the influence of different weather conditions

```{r}
top_weather_delay <- delay_departure_data %>%
  group_by(weather_description) %>%
  summarise(delay_mean = mean(delay_departure_min), delay_median = median(delay_departure_min), count = n()) %>%
  arrange(-delay_mean)

head(top_weather_delay, n=20)
```
We can also look at how temperature affects departure delays. The values of the temperature are always integers in the historical weather data I collected.

```{r}
average_delay_temperature <- delay_departure_data %>%
  group_by(temperature_celsius) %>%
  summarize(average_delay = mean(delay_departure_min, na.rm = TRUE), count = n())

ggplot(average_delay_temperature, aes(x = temperature_celsius, y = average_delay)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Temperature (Celsius)", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Temperature on Departure Delay") +
  coord_cartesian(ylim = c(0, 100))
```

If one looks at the low temperature, then a valley, which can be seen even more clearly in the graph above.

```{r}
average_delay_templow <- delay_departure_data %>%
  group_by(temperature_celsius_low) %>%
  summarize(average_delay = mean(delay_departure_min, na.rm = TRUE))

ggplot(average_delay_templow, aes(x = temperature_celsius_low, y = average_delay)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Low Temperature", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Low Temperature on Departure Delay")
```

```{r}
average_delay_wind <- delay_departure_data %>%
  group_by(wind_speed_km) %>%
  summarize(average_delay = mean(delay_departure_min, na.rm = TRUE))

ggplot(average_delay_wind, aes(x = wind_speed_km, y = average_delay)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Wind Speed (km/h)", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Wind Speed on Departure Delay") +
  coord_cartesian(ylim = c(0, 100))
```

```{r}
average_delay_wind_direction <- delay_departure_data %>%
  group_by(wind_direction) %>%
  summarize(average_delay = mean(delay_departure_min, na.rm = TRUE))

ggplot(average_delay_wind_direction, aes(x = wind_direction, y = average_delay)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Wind Direction", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Wind Direction on Departure Delay") +
  coord_cartesian(ylim = c(0, 30))
```

```{r}
average_delay_humidity <- delay_departure_data %>%
  group_by(humidity) %>%
  summarize(average_delay = mean(delay_departure_min, na.rm = TRUE))

ggplot(average_delay_humidity, aes(x = humidity, y = average_delay)) +
  geom_bar(stat = "identity", fill = "steelblue") +
  labs(x = "Humidity", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Humidity on Departure Delay")

```
Looking at barometric pressure, I can personally not see any interessting trends...

```{r}
average_delay_baro <- delay_departure_data %>%
  group_by(baro) %>%
  summarize(average_delay = mean(delay_departure_min, na.rm = TRUE))

ggplot(average_delay_baro, aes(x = baro, y = average_delay)) +
  geom_line(color = "steelblue") +
  labs(x = "Baro", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Baro on Departure Delay")
```

```{r}
ggplot(average_delay_baro, aes(x = baro, y = average_delay)) +
  geom_line(color = "steelblue") +
  labs(x = "Baro", y = "Average Departure Delay (Minutes)") +
  ggtitle("Influence of Baro on Departure Delay") +
  coord_cartesian(xlim = c(970, 1050))
```
```{r}
summary(average_delay_baro)
```
