==> 01 flight delay prediction/04_predict_departure_delay.Rmd <==
# Notebook: Predict Flight Delay

Now, the aim is to predict whether a particular flight has (a) little/no delay or (b) long/very long delay, i.e. a binary classification. Previously, I tried to predict the delay of flights departing between 05/2022 and 05/2023 using the flight data from 10/2017 to 05/2022, also in a binary way. I was able to achieve a very poor accuracy (about 50%). Previously, the data were divided into five different classes using the 0.2-, 0.4-, 0.6- and 0.8-quantiles (5 classes) whereby the accuracy was also only around 20%.

## Load Packages

```{r}
library(ggplot2)
library(dplyr)
library(stringr)
library(caret)
library(e1071)
source("get_weather_data.R")
```

## Settings

```{r}
set.seed(42)
model_name_method <- "svmLinear"
```

## Code 

### Load Datasets

First, we need to load the dataset and add some columns that will be used as features (`weather_data`)

```{r}  
# Load departure data
departure_data <- read.csv("../datasets/departure_dataset.csv")
weather_data <- read.csv("../datasets/weather_munich_dataset.csv")
head(departure_data)
```

### Convert to POSIXct

```{r}
departure_data$departure_scheduled_datetime_local <- as.POSIXct(departure_data$departure_scheduled_datetime_local)
weather_data$six_hours_starting_from_datatime <- as.POSIXct(paste(weather_data$date, weather_data$six_hours_starting_from))
```

### Add Weather Data

```{r}
departure_data$six_hours_starting_from_datatime <- as.POSIXct(sapply(departure_data$departure_scheduled_datetime_local, get_weather_data))
merge_cols_weather_data <- setdiff(names(weather_data), c("date", "six_hours_starting_from"))
departure_data <- merge(departure_data, 
                        weather_data[, c(merge_cols_weather_data)], 
                        by.x = "six_hours_starting_from_datatime", 
                        by.y = "six_hours_starting_from_datatime", 
                        all.x = TRUE)
head(departure_data)
```

### Remove Special Characters from Values

```{r}
# Since there should be a separate column for each value for weather_description later, which then has its value as the column name, I remove all "." and spaces so that you can access the column name later using $ in R.
departure_data$weather_description <- gsub("[. ]", "", departure_data$weather_description)
departure_data$airline_name <- gsub("[. ]", "", departure_data$airline_name)
departure_data$destination_IATA <- gsub("[. ]", "", departure_data$destination_IATA)
head(departure_data)
```

```{r}
# There are cases where no weather description is given but other weather data such as temperature is available. 
# In this case, I set a value for these cases to "NotSpecified" so that the value is not NA. Later, such cases get their own column for the one-hot encoding that will be applied.
departure_data$weather_description <- ifelse(is.na(departure_data$weather_description), "NotSpecified", departure_data$weather_description)
```

### Select Flights

As I want to predict the amount of delay, I will only look at flights that have landed (not canceled, unknown status etc.)

```{r}
departure_data <- departure_data[departure_data$flight_status == "Landed", ]
```

### Feature Coding

```{r}
dataset <- departure_data %>%
  select(departure_scheduled_datetime_local,
         delay_departure_min,
         departure_scheduled_weekday_local,
         departure_scheduled_hour_local, # continuous variable
         departure_scheduled_month_local, # continuous variable 
         departure_scheduled_year_local, # continuous variable
         temperature_celsius,
         temperature_celsius_low,
         baro,
         wind_speed_km,
         wind_direction,
         humidity)
```

```{r}
dataset$departure_scheduled_weekday_local <- as.integer(factor(dataset$departure_scheduled_weekday_local,
                                                                      levels = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"),
                                                                      labels = 1:7))
```

```{r}
# One-hot encoding for categorical variables (note for me: 2 means that columns)
dataset <- cbind(dataset, apply(model.matrix(~ weather_description - 1, data = departure_data), 2, as.integer)) 
dataset <- cbind(dataset, apply(model.matrix(~ airline_name - 1, data = departure_data), 2, as.integer))
dataset <- cbind(dataset, apply(model.matrix(~ destination_IATA - 1, data = departure_data), 2, as.integer))
```

```{r}
# Now there are one hot encoded columns
head(dataset[, c("weather_descriptionMostlycloudy", "weather_descriptionMorecloudsthansun", "weather_descriptionMild")])
```

### Remove Outliers

There are some extreme outliers in the dataset (related to departure delay). I have no explanation for this, probably some technical problems...

```{r}
head(dataset[order(dataset$delay_departure_min, decreasing = TRUE), ], 5)[, c("delay_departure_min", "departure_scheduled_datetime_local")]
```

```{r}
head(dataset[order(dataset$delay_departure_min), ], 5)[, c("delay_departure_min", "departure_scheduled_datetime_local")]
```

In order to remove outliers, the code calculates the z-scores based on the "delay_departure_min" column in the dataset using the scale() function. The z-score measures how many standard deviations an observation is away from the mean.

```{r}
z_scores <- scale(dataset$delay_departure_min)
z_threshold <- 3
outliers_train_df <- abs(z_scores) > z_threshold
dataset <- dataset[!outliers_train_df, ]
```

### Prepare List of Year and Months

The model should always be trained with flight data from one month and evaluated with flight data from the following month.

```{r}
year_month_pairs <- dataset %>%
  group_by(departure_scheduled_month_local, departure_scheduled_year_local) %>%
  summarise(average_delay = mean(delay_departure_min)) %>%
  arrange(departure_scheduled_year_local, departure_scheduled_month_local)
head(year_month_pairs)
```

### Define Classes

```{r}
category_names_quantiles <- c("earlier, on time or very short delay", "long delay")
category_quantiles <- quantile(dataset$delay_departure_min, probs = c(0.0, 0.5, 1.0))
category_quantiles
```

### Train Model

#### Method to Train Model

```{r}
train_model <- function (model_name="rf") {
  ctrl <- trainControl(verboseIter = TRUE, method="none")
  filtered_train_df <- train_df[-(which(names(train_df) %in% c("departure_scheduled_datetime_local", "delay_departure_min")))]
  
  model <- train(delay_departure_min_category ~ .,
         data = filtered_train_df,
         method = model_name,
         importance = TRUE,
         trControl = ctrl,
         )
  return(model)
}
```

#### Method to Evaluate Model

```{r}
evaluate_model <- function (model, test_df) {
  predictions <- predict(model, newdata = test_df)
  cf <- confusionMatrix(predictions, test_df$delay_departure_min_category)
  feature_importance <- varImp(model)
  
  # Now I extract the variable importance from all variables (in alphabetic order)
  feature_importance_row <- feature_importance$importance%>% 
    mutate(names=row.names(.)) %>%
    arrange(names) %>%
    t() %>%
    as.data.frame()
  
  new_row <- c(year_start_test = year_start_test, 
               month_start_test = month_start_test,
               year_start_train = year_start_train,
               month_start_train = month_start_train,
               cf$overall, 
               cf$byClass, 
               as.character(unlist(feature_importance_row[1, ]))
               ) 
  results_df <- rbind(results_df, new_row)
  colnames(results_df) <- c("year_start_test", "month_start_test", "year_start_train", "month_start_train", names(cf$overall), names(cf$byClass), colnames(feature_importance_row))
  return(results_df)
}
```

```{r}
results_df <- data.frame()
```

#### Train/Evaluate Model for each month

```{r}
for (i in seq(1, nrow(year_month_pairs) - 1)) {
  year_start_train <- year_month_pairs$departure_scheduled_year_local[[i]]
  year_start_test <- year_month_pairs$departure_scheduled_year_local[[i+1]]
  month_start_train <- year_month_pairs$departure_scheduled_month_local[[i]]
  month_start_test <- year_month_pairs$departure_scheduled_month_local[[i+1]]
  
  # Filter flight data for a two months, one for each training and testing
  train_df <- dataset %>%
    filter(departure_scheduled_year_local == year_start_train & departure_scheduled_month_local == month_start_train)
  test_df <- dataset %>%
    filter(departure_scheduled_year_local == year_start_test & departure_scheduled_month_local == month_start_test)

  
  # Add Labels
  train_df$delay_departure_min_category <- cut(train_df$delay_departure_min, 
                                             breaks = category_quantiles, 
                                             labels = category_names_quantiles,
                                             include.lowest = TRUE)
  
  test_df$delay_departure_min_category <- cut(test_df$delay_departure_min, 
                                             breaks = category_quantiles, 
                                             labels = category_names_quantiles,
                                             include.lowest = TRUE) 
  
  # Replace NA with average
  columns <- c("temperature_celsius", "temperature_celsius_low", "baro", "wind_speed_km", "wind_direction", "humidity")

  for (col in columns) {
     average_value_train <- mean(train_df[[col]], na.rm = TRUE)
     average_value_test <- mean(test_df[[col]], na.rm = TRUE)
  
     train_df[[col]] <- ifelse(is.na(train_df[[col]]), average_value_train, train_df[[col]])
     test_df[[col]] <- ifelse(is.na(test_df[[col]]), average_value_test, test_df[[col]])
  }

  # Train Model
  model <- train_model(model_name=model_name_method)
  
  # Save new results
  results_df <- evaluate_model(model, test_df)
}
```

```{r}
write.csv(results_df, paste0("results/results_", model_name_method, ".csv"))
head(results_df)
```
