==> 01 flight delay prediction/05_analyse_results.Rmd <==
# Notebook: Analyse Results of Trained Model

## Load Packages

```{r}
library(ggplot2)
library(dplyr)
library(tidyverse)
```

## Define Settings

```{r}
top_n <- 7
model_name <- "rf"
```

## Code

### Add Column

In order to better visualize the results of the trained model, I will add a column containing both year and month.

```{r}
evaluation_results <- read.csv(paste0("results/results_", model_name, ".csv"))

evaluation_results <- evaluation_results %>%
  mutate(year_month_train = paste(year_start_train, month_start_train, sep = "-"))

evaluation_results[, c("year_month_train", "Accuracy")]
```

Compared to the other years, the classification accuracy for the months in 2020 is above 80% in most months and thus very high. Here, it should be noted that in 2020, the delay of departure for the flights that took off from Munich were rather low, which is why the upper class (above the median of the overall dataset) was rather rare in comparison to the lower class (below median of the overall dataset). See Notebook 03_dataset_exploration, which shows that the amount of delay was lower in 2020 comparing it with the amount of delay in the other years.

```{r, fig.width = 10, fig.height= 3}
# Date with day is required for visualisation, only year and month can't be transformed into a value of type Date
evaluation_results$year_month_train_first_day <- as.Date(paste0(evaluation_results$year_month_train, "-01"))
x_labels_months <- unique(evaluation_results$year_month_train_first_day)

for (year in unique(year(evaluation_results$year_month_train_first_day))) {
  year_data <- subset(evaluation_results, year(evaluation_results$year_month_train_first_day) == year)
  plot <- ggplot(year_data, aes(x = year_month_train_first_day, y = Accuracy)) +
    geom_line() +
    labs(x = "Month", y = "Variable Accuracy", color = "Feature") +
    scale_x_date(breaks = x_labels_months, labels = format(x_labels_months, "%b %Y")) +
    ggtitle(paste("Year", year)) +
    ylim(0.4, 1)
  print(plot) 
}
```

### Calculate mean of each column

I want to calculate the mean of each metric based on all months for which I trained the model.

```{r}
metrics_df <- evaluation_results[, 6:23]

metrics <- data.frame(
  Mean = apply(metrics_df, 2, mean, na.rm = TRUE),
  Median = apply(metrics_df, 2, median, na.rm = TRUE),
  Min = apply(metrics_df, 2, min, na.rm = TRUE),
  Max = apply(metrics_df, 2, max, na.rm = TRUE),
  Q1 = apply(metrics_df, 2, quantile, probs = 0.25, na.rm = TRUE),
  Q3 = apply(metrics_df, 2, quantile, probs = 0.75, na.rm = TRUE)
)

row.names(metrics) <- colnames(metrics_df)
metrics
```

###  Get Most Important Features

Now, I want to select the features with the highest average variable importance. First of all, I select the columns with the features and it's variable importance scores.

```{r}
feature_columns <- evaluation_results[, 24:ncol(evaluation_results)]
head(feature_columns[, 1:3])
```

```{r}
column_means <- sapply(feature_columns, mean)
sorted_means <- sort(column_means, decreasing = TRUE)
top_n_columns <- names(sorted_means)[1:top_n]
head(top_n_columns)
```


### Format Data for Visualisation of the Development of the Feature Importance Over Time

```{r}
selected_feature_timeseries <- evaluation_results %>%
  select(all_of(top_n_columns), year_month_train_first_day)
selected_feature_timeseries <- tidyr::gather(selected_feature_timeseries, key = "feature", value = "importance", -year_month_train_first_day)
head(selected_feature_timeseries)
```

### Create Visualisation for each Year

To visualize the development of the feature importance over time, I for each year and month a dataframe (`year_data`) with three columns is created.

Columns:

* the month and year of the test data that should be evaluated
* the name of the feature
* the importance of the feature.

```{r, fig.width = 10, fig.height= 3}
for (year in unique(year(selected_feature_timeseries$year_month_train_first_day))) {
  year_data <- subset(selected_feature_timeseries, year(selected_feature_timeseries$year_month_train_first_day) == year)
  plot <- ggplot(year_data, aes(x = year_month_train_first_day, y = importance, color = feature)) +
    geom_line() +
    labs(x = "Month", y = "Variable Importance", color = "Feature") +
    scale_x_date(breaks = x_labels_months, labels = format(x_labels_months, "%b %Y")) +
    ggtitle(paste("Year", year))
  
  filename <- paste0("plots/", "plot_", model_name, "_", year, ".png", sep = "")
  ggsave(filename, plot = plot, width = 20, height = 10, dpi = 300)
  print(plot) 
}
```
